+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Data cleaning for USA
 - Data cleaning for Japan
 - Data cleaning for AUS
+- Merge data from different countries
 
 
 # FeliX3_YoGL_v25 (by Q. Ye)